legacy/backend/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.44.0
google-generativeai==0.8.6
pydantic==2.12.5
python-dotenv==1.2.2
requests==2.33.1
